requirements.txt
Django
psycopg2
django-daterange-filter
django-import-export
